Try to Make Player Spawn Deterministic
Sort the map spawners by their order in the map,
so that the spawner selection is deterministic.
